+++ include/boost/geometry/strategies/cartesian/intersection.hpp
