+++ include/boost/geometry/strategies/geographic/azimuth.hpp
