+++ include/boost/geometry/strategies/intersection_strategies.hpp
@@ -22,10 +22,10 @@
 #include <boost/geometry/policies/relate/tupled.hpp>
 
 #include <boost/geometry/strategies/intersection.hpp>
-#include <boost/geometry/strategies/side.hpp>
 #include <boost/geometry/strategies/intersection_result.hpp>
+#include <boost/geometry/strategies/side.hpp>
 
-#include <boost/geometry/strategies/cartesian/cart_intersect.hpp>
+#include <boost/geometry/strategies/cartesian/intersection.hpp>
 #include <boost/geometry/strategies/cartesian/side_by_triangle.hpp>
 #include <boost/geometry/strategies/spherical/intersection.hpp>
 #include <boost/geometry/strategies/spherical/ssf.hpp>
